+++ CHANGELOG.md
@@ -2,12 +2,11 @@
 
 * Add support for separate database
 * Add basic tasks for cleaning up
-* Improve index by adding a search
 * Run tests on different ruby/rails/database combinations
 
 # 0.1.0
 
 * Added an interceptor and database table
-* Added views to visualize emails
+* Added views to visualize emails and search them
 * Added delivery method to open with Launchy